pyproject.toml
[project]
name = "photo-merger"
version = "0.1.0"
description = "Merge photos from multiple directories chronologically into a single directory by extracting metdata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "logging>=0.4.9.6",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.11.6",
]


[[tool.mypy.overrides]]
module = ["yaml.*"]
follow_untyped_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
